base_db

Module input

This module specifies the input to rust-analyzer. In some sense, this is the most important module, because all other fancy stuff is strictly derived from this input.

Note that neither this module, nor any other part of the analyzer’s core do actual IO. See vfs and project_model in the rust-analyzer crate for how actual IO is done and lowered to input.

Structs

Crate

CrateBuilder

CrateData

CrateDisplayName

CrateGraphBuilder

CrateName

CratesMap

The mapping from UniqueCrateData to their Crate input.

CyclicDependenciesError

Dependency

Env

ExtraCrateData

Crate data unrelated to analysis.

SourceRoot

Files are grouped into source roots. A source root is a directory on the file systems which is watched for changes. Typically it corresponds to a Rust crate. Source roots might be nested: in this case, a file belongs to the nearest enclosing source root. Paths to files are always relative to a source root, and the analyzer does not know the root path of the source root at all. So, a file from one source root can’t refer to a file in another source root by path.

SourceRootId

UniqueCrateData

The crate data from which we derive the Crate.

Enums

CrateOrigin

Origin of the crates.

LangCrateOrigin

ProcMacroLoadingError

ReleaseChannel

Type Aliases

BuiltCrateData

BuiltDependency

CrateBuilderId

CrateDataBuilder

CratesIdMap

DependencyBuilder

ProcMacroPaths