base_db::salsa

Enum EventKind

pub enum EventKind {

    DidValidateMemoizedValue {
        database_key: DatabaseKeyIndex,
    },
    WillBlockOn {
        other_thread_id: ThreadId,
        database_key: DatabaseKeyIndex,
    },
    WillExecute {
        database_key: DatabaseKeyIndex,
    },
    WillIterateCycle {
        database_key: DatabaseKeyIndex,
        iteration: u8,
    },
    DidFinalizeCycle {
        database_key: DatabaseKeyIndex,
        iteration: u8,
    },
    WillCheckCancellation,
    DidSetCancellationFlag,
    WillDiscardStaleOutput {
        execute_key: DatabaseKeyIndex,
        output_key: DatabaseKeyIndex,
    },
    DidDiscard {
        key: DatabaseKeyIndex,
    },
    DidDiscardAccumulated {
        executor_key: DatabaseKeyIndex,
        accumulator: DatabaseKeyIndex,
    },
    DidInternValue {
        key: DatabaseKeyIndex,
        revision: Revision,
    },
    DidReuseInternedValue {
        key: DatabaseKeyIndex,
        revision: Revision,
    },
    DidValidateInternedValue {
        key: DatabaseKeyIndex,
        revision: Revision,
    },

}

An enum identifying the various kinds of events that can occur.

Variants

DidValidateMemoizedValue

Occurs when we found that all inputs to a memoized value are up-to-date and hence the value can be re-used without executing the closure.

Executes before the “re-used” value is returned.

Fields

database_key: DatabaseKeyIndex

The database-key for the affected value. Implements Debug.

WillBlockOn

Indicates that another thread (with id other_thread_id) is processing the given query (database_key), so we will block until they finish.

Executes after we have registered with the other thread but before they have answered us.

Fields

other_thread_id: ThreadId

The id of the thread we will block on.

database_key: DatabaseKeyIndex

The database-key for the affected value. Implements Debug.

WillExecute

Indicates that the function for this query will be executed. This is either because it has never executed before or because its inputs may be out of date.

Fields

database_key: DatabaseKeyIndex

The database-key for the affected value. Implements Debug.

WillIterateCycle

Salsa starts a new fixpoint iteration for the cycle with database_key as its outermost cycle.

Fields

database_key: DatabaseKeyIndex

The database-key for the cycle head. Implements Debug.

iteration: u8

DidFinalizeCycle

Salsa completed a fixpoint iteration for the cycle with database_key as its outermost cycle.

Fields

database_key: DatabaseKeyIndex

iteration: u8

WillCheckCancellation

Indicates that unwind_if_cancelled was called and salsa will check if the current revision has been cancelled.

DidSetCancellationFlag

Indicates that one Handle has set the cancellation flag. When other active handles execute salsa methods, they will observe this flag and panic with a sentinel value of type Cancelled.

WillDiscardStaleOutput

Discovered that a query used to output a given output but no longer does.

Fields

execute_key: DatabaseKeyIndex

Key for the query that is executing and which no longer outputs the given value.

output_key: DatabaseKeyIndex

Key for the query that is no longer output

DidDiscard

Tracked structs or memoized data were discarded (freed).

Fields

key: DatabaseKeyIndex

Value being discarded.

DidDiscardAccumulated

Discarded accumulated data from a given fn

Fields

executor_key: DatabaseKeyIndex

The key of the fn that accumulated results

accumulator: DatabaseKeyIndex

Accumulator that was accumulated into

DidInternValue

Indicates that a value was newly interned.

Fields

key: DatabaseKeyIndex

revision: Revision

DidReuseInternedValue

Indicates that a value was interned by reusing an existing slot.

Fields

key: DatabaseKeyIndex

revision: Revision

DidValidateInternedValue

Indicates that a previously interned value was read in a new revision.

Fields

key: DatabaseKeyIndex

revision: Revision