Module base_db::span

File and span related types.

Structs

HirFileId

Input to the analyzer is a set of files, where each file is identified by FileId and contains source code. However, another source of source code in Rust are macros: each macro can be thought of as producing a “temporary file”. To assign an id to such a file, we use the id of the macro call that produced the file. So, a HirFileId is either a FileId (source code written by user), or a MacroCallId (source code produced by macro).

MacroCallId

MacroCallId identifies a particular macro invocation, like println!("Hello, {}", world).

MacroFileId

SpanAnchor

SyntaxContextId

Enums

HirFileIdRepr

Constants

ROOT_ERASED_FILE_AST_ID

Type Aliases

ErasedFileAstId

SpanData