base_db

Type Alias CrateDataBuilder

pub type CrateDataBuilder = CrateData<CrateBuilderId>;

Aliased Type

pub struct CrateDataBuilder {
    pub root_file_id: FileId,
    pub edition: Edition,
    pub dependencies: Vec<Dependency<Idx<CrateBuilder>>>,
    pub origin: CrateOrigin,
    pub is_proc_macro: bool,
    pub proc_macro_cwd: Arc<AbsPathBuf>,
}

Fields
root_file_id: FileIdedition: Editiondependencies: Vec<Dependency<Idx<CrateBuilder>>>

The dependencies of this crate.

Note that this may contain more dependencies than the crate actually uses. A common example is the test crate which is included but only actually is active when declared in source via extern crate test.

origin: CrateOriginis_proc_macro: boolproc_macro_cwd: Arc<AbsPathBuf>

The working directory to run proc-macros in invoked in the context of this crate. This is the workspace root of the cargo workspace for workspace members, the crate manifest dir otherwise.