Crate base_db

base_db defines basic database traits. The concrete DB is defined by ide.

Re-exports

pub use ra_salsa;

Macros

impl_intern_key

Structs

AnchoredPath

Path relative to a file.

AnchoredPathBuf

Path relative to a file.

BuildMetadata

Optional build metadata identifier. This comes after + in a SemVer version, as in 0.8.1+zstd.1.5.0.

CompressedFileTextQuery

CrateData

CrateDisplayName

CrateGraph

CrateGraph is a bit of information which turns a set of text files into a number of Rust crates.

CrateGraphQuery

CrateName

CrateWorkspaceData

Crate related data shared by the whole workspace.

CrateWorkspaceDataQuery

Dependency

Env

FileChange

Encapsulate a bunch of raw .set calls on the database.

FileLoaderDelegate

Silly workaround for cyclic deps due to the SourceRootDatabase and SourceDatabase split regarding FileLoader

FileSet

A set of VfsPaths identified by [FileId]s.

FileSourceRootQuery

FileTextQuery

ParseErrorsQuery

ParseQuery

Prerelease

Optional pre-release identifier on a version string. This comes after - in a SemVer version, like 1.0.0-alpha.1

SourceDatabaseGroupStorage__

SourceDatabaseStorage

Representative struct for the query group.

SourceRoot

Files are grouped into source roots. A source root is a directory on the file systems which is watched for changes. Typically it corresponds to a Rust crate. Source roots might be nested: in this case, a file belongs to the nearest enclosing source root. Paths to files are always relative to a source root, and the analyzer does not know the root path of the source root at all. So, a file from one source root can’t refer to a file in another source root by path.

SourceRootCratesQuery

SourceRootDatabaseGroupStorage__

SourceRootDatabaseStorage

Representative struct for the query group.

SourceRootId

SourceRootQuery

Version

SemVer version as defined by https://semver.org.

VersionReq

SemVer version requirement describing the intersection of some version comparators, such as >=1.2.3, <1.8.

VfsPath

Path in Vfs.

Enums

Cancelled

A panic payload indicating that execution of a salsa query was cancelled.

CrateOrigin

Origin of the crates.

LangCrateOrigin

ReleaseChannel

Constants

DEFAULT_BORROWCK_LRU_CAP

DEFAULT_FILE_TEXT_LRU_CAP

DEFAULT_PARSE_LRU_CAP

Traits

FileLoader

SourceDatabase

Database which stores all significant input facts: source code and project model. Everything else in rust-analyzer is derived from these queries.

SourceDatabaseFileInputExt

SourceRootDatabase

We don’t want to give HIR knowledge of source roots, hence we extract these methods into a separate DB.

Upcast

Type Aliases

CrateId

ProcMacroPaths

TargetLayoutLoadResult