Crate base_db

base_db defines basic database traits. The concrete DB is defined by ide.

Re-exports

pub use salsa;

pub use salsa_macros;

pub use query_group;

Modules

target

Information about the target.

Macros

impl_intern_key

Structs

AnchoredPath

Path relative to a file.

AnchoredPathBuf

Path relative to a file.

BuildMetadata

Optional build metadata identifier. This comes after + in a SemVer version, as in 0.8.1+zstd.1.5.0.

Crate

CrateBuilder

CrateDisplayName

CrateGraphBuilder

CrateName

CrateWorkspaceData

Crate related data shared by the whole workspace.

CratesMap

The mapping from UniqueCrateData to their Crate input.

DbPanicContext

EditionedFileId

Env

ExtraCrateData

Crate data unrelated to analysis.

FileChange

Encapsulate a bunch of raw .set calls on the database.

FileId

Handle to a file in [Vfs]

FileSet

A set of VfsPaths identified by FileIds.

FileSourceRootInput

FileText

Files

Nonce

Prerelease

Optional pre-release identifier on a version string. This comes after - in a SemVer version, like 1.0.0-alpha.1

SourceRoot

Files are grouped into source roots. A source root is a directory on the file systems which is watched for changes. Typically it corresponds to a Rust crate. Source roots might be nested: in this case, a file belongs to the nearest enclosing source root. Paths to files are always relative to a source root, and the analyzer does not know the root path of the source root at all. So, a file from one source root can’t refer to a file in another source root by path.

SourceRootId

SourceRootInput

UniqueCrateData

The crate data from which we derive the Crate.

Version

SemVer version as defined by https://semver.org.

VersionReq

SemVer version requirement describing the intersection of some version comparators, such as >=1.2.3, <1.8.

VfsPath

Path in Vfs.

Enums

CrateOrigin

Origin of the crates.

LangCrateOrigin

ProcMacroLoadingError

ReleaseChannel

Constants

DEFAULT_BORROWCK_LRU_CAP

DEFAULT_FILE_TEXT_LRU_CAP

DEFAULT_PARSE_LRU_CAP

Traits

RootQueryDb

Database which stores all significant input facts: source code and project model. Everything else in rust-analyzer is derived from these queries.

SourceDatabase

Type Aliases

BuiltCrateData

BuiltDependency

CrateBuilderId

CrateDataBuilder

CratesIdMap

DependencyBuilder

FxIndexMap

FxIndexSet

ProcMacroPaths