Trait base_db::SourceDatabase

pub trait SourceDatabase: Database + HasQueryGroup<SourceDatabaseStorage> + CheckCanceled + Debug {
    fn parse(&self, key0: FileId) -> Parse<SourceFile>;

    fn file_text(&self, key0: FileId) -> Arc<String>;

    fn set_file_text(&mut self, key0: FileId, value__: Arc<String>);

    fn set_file_text_with_durability(
        &mut self, 
        key0: FileId, 
        value__: Arc<String>, 
        durability__: Durability
    );
}

Database which stores all significant input facts: source code and project model. Everything else in rust-analyzer is derived from these queries.

Required methods

fn parse(&self, key0: FileId) -> Parse<SourceFile>

fn file_text(&self, key0: FileId) -> Arc<String>

fn set_file_text(&mut self, key0: FileId, value__: Arc<String>)

Set the value of the file_text input.

See file_text for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_file_text_with_durability(
&mut self,
key0: FileId,
value__: Arc<String>,
durability__: Durability
)

Set the value of the file_text input and promise that its value will never change again.

See file_text for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

Loading content...